Move scheduler
